requirements.txt
aioapns==4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
firebase-admin==6.0.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h2==4.2.0
hpack==4.1.0
http-ece==1.2.1
httplib2==0.22.0
hyper==0.7.0
hyperframe==6.1.0
idna==3.10
msgpack==1.1.0
multidict==6.4.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dotenv==1.1.0
pywebpush==2.0.3
redis==5.2.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.0
